+++ src/views/system-tools/Raw-Data/RawData.tsx
@@ -572,7 +572,7 @@ const RowDataListing = () => {
       header: "Updated Info", accessorKey: "updated_at", enableSorting: true, size: 170, meta: { HeaderClass: "text-red-500" },// Matched size with Domain module
       cell: (props) => {
         const { updated_at, updated_by_name, updated_by_role } = props.row.original;
-        const formattedDate = updated_at ? `${new Date(updated_at).getDate()} ${new Date(updated_at).toLocaleString("en-US", { month: "long" })} ${new Date(updated_at).getFullYear()}, ${new Date(updated_at).toLocaleTimeString("en-US", { hour: "numeric", minute: "2-digit", hour12: true })}` : "N/A";
+        const formattedDate = updated_at ? `${new Date(updated_at).getDate()} ${new Date(updated_at).toLocaleString("en-US", { month: "short" })} ${new Date(updated_at).getFullYear()}, ${new Date(updated_at).toLocaleTimeString("en-US", { hour: "numeric", minute: "2-digit", hour12: true })}` : "N/A";
         return (<div className="text-xs"><span>{updated_by_name || "N/A"}{updated_by_role && (<><br /><b>{updated_by_role}</b></>)}</span><br /><span>{formattedDate}</span></div>);
       },
     },
