+++ src/views/web-settings/Trending-Carousel/TrendingCarousel.tsx
@@ -584,7 +584,7 @@ const TrendingCarousel = () => {
         cell: (props) => {
             const { updated_at, updated_by_name, updated_by_role } = props.row.original;
             const formattedDate = updated_at
-            ? `${new Date(updated_at).getDate()} ${new Date(updated_at).toLocaleString("en-US", { month: "long" })} ${new Date(updated_at).getFullYear()}, ${new Date(updated_at).toLocaleTimeString("en-US", { hour: "numeric", minute: "2-digit", hour12: true })}`
+            ? `${new Date(updated_at).getDate()} ${new Date(updated_at).toLocaleString("en-US", { month: "short" })} ${new Date(updated_at).getFullYear()}, ${new Date(updated_at).toLocaleTimeString("en-US", { hour: "numeric", minute: "2-digit", hour12: true })}`
             : "N/A";
             return (
                 <div className="text-xs">
